formally replace old reducers with new ones
